+++ server/src/modules/pages/db/entity/UrlPageIdEntity.ts
@@ -1,0 +1,4 @@
+export interface UrlPageIdEntity {
+  websiteId: string;
+  pageId: string;
+}
