+++ server/src/modules/pages/db/entity/index.ts
@@ -1,3 +1,4 @@
 export type { PageEntity } from "./PageEntity";
 export type { PageConfigEntity } from "./PageConfigEntity";
 export type { PageDetailsEntity } from "./PageDetailsEntity";
+export type { UrlPageIdEntity } from "./UrlPageIdEntity";
