+++ server/src/modules/pages/db/index.ts
@@ -1,6 +1,13 @@
-export type { PageEntity, PageConfigEntity, PageDetailsEntity } from "./entity";
+export type {
+  PageEntity,
+  PageConfigEntity,
+  PageDetailsEntity,
+  UrlPageIdEntity,
+} from "./entity";
 
 export type { UrlMeta, PageDetailsData } from "./page";
 export { PageRepository } from "./page";
+
+export { PageRankingRepository } from "./page-ranking";
 
 export { PageConfigRepository } from "./page-config";
