+++ server/src/modules/pages/db/page-ranking/aggregations/TopRatedPagesAggregation.ts
@@ -1,0 +1,17 @@
+export interface TopRatedPagesAggregation {
+  _id: {
+    websiteId: string;
+    pageId: string;
+  };
+  normalizedUrl: string;
+  votesSum: number;
+  votesUp: string[] | null;
+  votesDown: string[] | null;
+  meta: null | {
+    logo: string | null;
+    title: string | null;
+  };
+  comments: {
+    count: number;
+  };
+}
