+++ server/src/modules/pages/db/page-ranking/aggregations/index.ts
@@ -1,0 +1,2 @@
+export type { TopCommentedPagesAggregation } from "./TopCommentedPagesAggregation";
+export type { TopRatedPagesAggregation } from "./TopRatedPagesAggregation";
